+++ front-end/package.json
@@ -18,6 +18,7 @@
         "@types/react-dom": "^18.0.9",
         "@types/react-redux": "^7.1.24",
         "@types/react-router-dom": "^5.3.3",
+        "axios": "^1.2.2",
         "framer-motion": "^6.5.1",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
