+++ front-end/src/App.tsx
@@ -4,8 +4,8 @@ import { Route, Routes } from "react-router-dom";
 import PlacementsModule from "./placements_module";
 import ResourcesModule from "./resources_module";
 import StaticModule from "./static_module"; 
-import {Login} from "./common/Access/Login";
-import {SignUp} from "./common/Access/SignUp";
+import {Login} from "./common/Auth/Login";
+import {SignUp} from "./common/Auth/SignUp";
 function App() {
     return (
         <>
