+++ front-end/src/common/Auth/Login.tsx
@@ -1,13 +1,9 @@
 import { Button, FormControl, FormLabel, Input } from "@chakra-ui/react";
 import React, { useState } from "react";
+import { login_api } from "../API/Access";
 
-export const SignUp = () => {
-    const [UserInfo, setUserInfo] = useState({
-        name: "",
-        email: "",
-        mobile_num: "",
-        password: ""
-    });
+export const Login = () => {
+    const [UserInfo, setUserInfo] = useState({ email: "", password: "" });
 
     const handlechange = (e: any) => {
         setUserInfo({ ...UserInfo, [e.target.type]: e.target.value });
@@ -15,25 +11,16 @@ export const SignUp = () => {
 
     const handleSubmit = (e: any) => {
         e.preventDefault();
-        console.log(UserInfo);
+        
+        const res=login_api(UserInfo.email, UserInfo.password);
+        console.log(res);
     };
 
     return (
         <div>
-            
             <FormControl>
-                <FormLabel>Name</FormLabel>
-                <Input type="name" onChange={handlechange} />
-            </FormControl>
-
-            <FormControl>
-                <FormLabel>Email Address</FormLabel>
+                <FormLabel>Email address</FormLabel>
                 <Input type="email" onChange={handlechange} />
-            </FormControl>
-
-            <FormControl>
-                <FormLabel>Mobile Number</FormLabel>
-                <Input type="mobile_num" onChange={handlechange} />
             </FormControl>
 
             <FormControl>
